+++ lib/devise_google_authenticatable/controllers/helpers.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module DeviseGoogleAuthenticator
-  module Controllers # :nodoc:
-    module Helpers # :nodoc:
+  module Controllers
+    module Helpers
       extend ActiveSupport::Concern
 
       included do