+++ test/lib/devise_google_authenticatable/controllers/helpers_test.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'test_helper'
 require 'devise_google_authenticatable/controllers/helpers'
 
@@ -13,21 +15,37 @@ class HelpersTest < ActiveSupport::TestCase
   end
 
   test 'can get otpauth_user' do
-    assert_equal "username@app", otpauth_user('username', 'app')
+    assert_equal 'username@app', otpauth_user('username', 'app')
   end
 
   test 'can get otpauth_user with a qualifier' do
-    assert_equal "username@app-qualifier", otpauth_user('username', 'app', '-qualifier')
+    assert_equal 'username@app-qualifier', otpauth_user('username', 'app', '-qualifier')
   end
 
   # fake image tag
-  def image_tag(src, *args)
+  def image_tag(src, *_)
     src
   end
+
+  def qrcode_png(qualifier: nil, issuer: nil)
+    build_qrcode_from(
+      build_qrcode_data_from(
+        'helpers_test',
+        'RailsApp',
+        @user.gauth_secret,
+        qualifier,
+        issuer
+      )
+    )
+  end
+
   test 'generate qrcode' do
-    assert_equal "https://chart.googleapis.com/chart?chs=200x200&chld=M|0&cht=qr&chl=otpauth%3A%2F%2Ftotp%2Fhelpers_test%40RailsApp%3Fsecret%3D", google_authenticator_qrcode(@user)
-    assert_equal "https://chart.googleapis.com/chart?chs=200x200&chld=M|0&cht=qr&chl=otpauth%3A%2F%2Ftotp%2Fhelpers_test%40RailsAppMyQualifier%3Fsecret%3D", google_authenticator_qrcode(@user, 'MyQualifier')
-    assert_equal "https://chart.googleapis.com/chart?chs=200x200&chld=M|0&cht=qr&chl=otpauth%3A%2F%2Ftotp%2Fhelpers_test%40RailsApp%3Fsecret%3D%26issuer%3DMyIssuer", google_authenticator_qrcode(@user, nil, 'MyIssuer')
-    assert_equal "https://chart.googleapis.com/chart?chs=200x200&chld=M|0&cht=qr&chl=otpauth%3A%2F%2Ftotp%2Fhelpers_test%40RailsAppMyQualifier%3Fsecret%3D%26issuer%3DMyIssuer", google_authenticator_qrcode(@user, 'MyQualifier', 'MyIssuer')
+    assert_equal qrcode_png, google_authenticator_qrcode(@user)
+    assert_equal qrcode_png(qualifier: 'MyQualifier'),
+                 google_authenticator_qrcode(@user, 'MyQualifier')
+    assert_equal qrcode_png(issuer: 'MyIssuer'),
+                 google_authenticator_qrcode(@user, nil, 'MyIssuer')
+    assert_equal qrcode_png(qualifier: 'MyQualifier', issuer: 'MyIssuer'),
+                 google_authenticator_qrcode(@user, 'MyQualifier', 'MyIssuer')
   end
 end